log_router_agent/requirements.txt
fastapi
uvicorn[standard]
google-cloud-pubsub
google-cloud-bigquery
google-cloud-logging
protobuf
httpx